Stone giants (the lowest-level H) are now just called giants.
The generic "giant" monster only existed so its corpse could be used
for things like giant mummies.  So I removed it and renamed the "stone
giant" to simply "giant", which allows it to be used in that capacity
now, as well as actually showing up in the game, giving us dual use
out of one monster -- we can use both the monster AND its corpse.