Sokoban rules conduct, livelogged historical event on completion.
This adds a one-time historical event the first time you enter the
Sokoban zoo, which (barring wizporting and such) proves that you have
completed the puzzles (either by solving them or by cheating).

Whether you followed the rules is tracked as a conduct.  Both
completion of Sokoban and violating the rules are reported in the
conducts screen (and at game end).  Keeping the rules is only
mentioned if you also completed Sokoban, since it's trivial to keep
the rules if you just don't bother to do Sokoban at all.

Once you have entered the Sokoban zoo, the puzzles are considered
complete, so the rules no longer apply (not for conduct and not
for guilt/luck purposes either).

All of this required adding a history_search function, so I did.