Consolidate meaningless variety-in-name-only (monsters, objects).
Do away with about two thirds of the different kinds of polearms.
The remaining ones are partisan, lucern hammer, glaive, halberd,
and of course the lance.  (The dwarvish mattock, which is listed
with the polearms in objects.c, has not been changed.  I don't
think most players think of it as a polearm, since it doesn't
use polearm skill or the polearm interface.)  If you were a big
fan of the bill-guerrichaurde, I'm sorry.  #name your halberd.

Several fruits have had their probabilities adjusted, because
"melon" is now the default name for custom fruit, edging out the
unapetizing "slime mold"; melon is no longer a separate fruit
otherwise.  (The source code still refers to them as SLIME_MOLD.)

The stiletto -- basically just a fancy word for knife with zero
game significance -- is also gone.

Also, carnivorous apes have been removed.  Their only really
special features were having a grabbing attack and being on the
list of "summon nasties" monsters, but there are several other
monsters with grabbing attacks on that list.