Rescue hezrou from the boring pile with a foul-stench passive attack.
Amazing what a little web search can turn up.  Apparently in the
source material, when hezrou are involved in melee, their skin emits a
foul-smelling liquid that has various negative effects on everything
within a radius.  Nausea (like in Brogue) would probably be ideal for
this, but eh, I went with stinking cloud, that'll do for now.  The
stinking cloud effect doesn't get used at all in the average game of
NetHack, so this takes an interesting mechanic that is already
implemented and makes it relevant, which I view as a good thing,
as well as giving hezrou an excuse to exist in the game.

Unfortunately, a similar search for vrock turned up some interesting
drawings but nothing that would really translate into a game-play
effect, so they're still in the boring bin for now.