Don't blame cramping when engraving fails due to status effects.
If you are blind, confused, stunned, or hallucinating, there
is a high probability that engraving will fail to write what
you intended.  This is unchanged from vanilla.  However, the
logic for the "cramping" message was assuming that any failure
was due to writing too much.

This commit introduces a new message for the case where engraving
fails due to status effects.  The wording is vague enough to allow the
same message to be used for all four relevant status effects, to avoid
needing to alter the logic to keep track of which thing actually
caused the failure.  (We'd need to nest if/else some five levels deep.
It's not worth it.)  If you have any of the four relevant status
effects when engraving fails, you get a message warning that it's
difficult to write in your current condition.  If you don't have any
status effects (likely the more usual case), then you get the cramping
message, as before.