Scale back a couple of balance changes that were too strong.
The big thing here is the base death drop probability.  I had
increased it, to compensate for the fact that death drops fall off
later in the game as you kill more things.  That was needed up to a
point, but I overdid it.  This scales it back a good bit, but it's
still true that the base rate is higher than in vanilla and that
the rate falls off as you kill a lot of monsters.

Also, unicorn horns now have somewhat improved durability, to stop
players from being terrified of ever using them.