Apply Dev Team fix, C343-325, nethack.rankin, Fri Aug 3 01:05:50 2007 +0000
A player had reported that melting ice in the Valkyrie quest and
filling the resulting pool with a boulder resulted in the pool being
described as a "moat".

The fix is to extend the level compiler with an extra per-level flag
that says, basically, "ice on this level is to be interpreted as
frozen pools, rather than frozen moat".

There is a side-effect:  unlike moat, pools can be evaporated away
with fire.  However, the existing water on the Valkyrie quest was
pools and thus already subject to this, so it's consistent.

Not applied:  the Dev Team patch also makes one of the drawbridges on
the final Valkyrie quest level start in a random position.  I did not
apply that change at this time, partly because it seems like 4.4
material, and partly because ultimately, if we're going to do that, it
should be random which of the two drawbridges might be closed.

Also not applied:  there was a second patch, nethack.rankin,
Fri Aug 3 01:09:13 2007 +0000, which made a whole bunch of totally
incomprehensible changes to sys/share/lev_yacc.c, with a commit
log comment to the effect that adding an extra parse rule had
changed yacc's tables.  We don't seem to have that file, so I
am guessing it is probably generated by yacc at compile time.
Then again, we don't have pretty much the entire sys/ tree, so
perhaps it's just fundamentally inapplicable to NH4.  Either way,
I did not attempt to apply that second commit.