Ticket #589, C343-265, apply Dev Team fix, nethack.rankin, Sun Feb 18 05:54:15 2007 +0000
If you are engulfed or for any other reason can't reach the floor,
don't allow forcing a chest lock.

Not applied:  the Dev Team's commit also contains a refinement to an
earlier fix having to do with the billing for breaking a lock on a
shop-owned chest.  I did not attempt to apply that portion, because in
general billing-related changes from the Dev Team's codebase seldom
apply cleanly to NH4.