More variety of maze design in Gehennom.
Thirty consecutive levels of mazes is boring no matter how you slice
it, but it's *really* boring when the mazes are all exactly the same
design.  This commit attempts to change things up just a little.

First, I cherry-picked the Gehennom Fun Patch change that allows mazes
to have corridor widths and/or wall thicknesses more than 1 tile wide.

Some sizes are prone to disconnected areas around the edges,
especially at the top and left.  I have not entirely solved this; I
suspect it may be partly a consequence of how mazexy() works, which I
did not change, but the problem does not manifest when both wall and
corridor thickness are 1.  Nonetheless, by the time the player reaches
Gehennom, he should have ways of dealing with disconnected areas.
They're not deliberate, and if I can figure out how to eliminate them
I will, but they aren't going to block the commit.

Second, occasionally there can be a maze level where the walls are
replaced by lava or, even more rarely, by iron bars or by graves.
The frequency of these possibilities may need to be tweaked.

I intend additional changes in Gehennom -- some levels that are not in
fact mazes at all, for example -- but this is a starting point.