Set monster's whip as not wielded before extracting it from his inventory.
This is an addendum to the fix for what happens when a monster
bullwhips your wielded in-use leash.  ais523 says it's safest to
consider obj_extract_self the first half of the drop/pickup operation,
an operatin that should not be even started while the object is
equipped.  Makes sense to me.

Basic testing does not turn up any player-noticeable difference
with this adjustment.