Apply Dev Team fix, C343-264, nethack.rankin, Sun Feb 18 03:33:43 2007 +0000
Improve the specificity of the "You must drop your weapon(s)" message,
and also only issue it if a weapon actually gets dropped.

Not applied:  the fix also includes improved (message) handling for
wielded tins and corpses, in the weapon_descr function in weapon.c,
but we do not currently have that function.  (It was introduced in
the fix for C343-60, which I have not yet applied.)