Apply Dev Team fix, UNL-052, nethack.rankin, Tue Apr 3 03:20:27 2007 +0000
In 3.4.3, monsters that can be generated with boulders in inventory
are no less likely to be generated in Sokoban than any other monster.
A player had suggested they should not be generated at all or perhaps
without inventory, but the fix does not go that far:  it just rejects
them on the first pass, making them quite a bit less likely.