Apply (part of) Dev Team fix, C343-267, nethack.rankin, Fri Feb 23 03:13:37 2007 +0000
A player had reported that zapping speed or slow monster at a sessile
monster produced a message to the effect that it was moving faster or
slower, although in fact it could not move at all.

The fix addresses permanently-sessile monsters, as well as those who
currently cannot move due to sleeping or paralysis.

Not applied:  the Dev Team's patch also adjustes the rules for when an
object becomes known as a result of changing a monster's speed, but it
used the learnwand() function, which NH4 does not have, so I did not
attempt to apply that portion.