Allow explore and setseed bones (in their own respective pools).
Now that all bones are in per-mode pools, we can allow them for all
modes of play.  Explore and setseed bones, like wizmode bones, will
not be seen in normal-mode play.

I'm doing this as a separate commit in case someone has additional,
non-normal-mode-related reasons that I can't think of right now to
disallow explore bones.