Move the Big Room a little deeper into the dungeon.
I've never liked the fact that you can get the Sokoban entrance
in the Big Room, which can make Sokoban quite tricky to reach,
especially for new players who particularly need it.

This commit moves the Big Room a bit lower down in the dungeon,
into a portion of the dungeon that needed a bit of tedium relief
in any case.