Apply Dev Team fix, C343-273, nethack.rankin, Sat Mar 10 05:54:17 2007 +0000
It had been reported that trying to invoke a wielded artifact after
changing alignment resulted in a message to the effect that the
artifact evaded your grasp, when in fact you were left wielding it.

The fix does two things.  First, it adjusts the message in
touch_artifact if the object is already in inventory.  Second, a new
function is added, retouch_object(), which handles cases wherein the
inability to touch an object ought to cause unwielding/unwearing.