Preliminary implementation of the ice devils' encase-in-iceblock attack.
When an ice devil attacks you, one of the things it can do is encase
you in a solid block of ice, temporarily immobilizing you.  This wears
off after a few turns, or there are various ways to escape it early
(e.g., melt the ice, dig out, teleport out).

This is highly experimental, and in particular I'm not sure the uhitm
case works entirely as intended.