Count boulders and pits when constructing a Sokoban .des from text input.
This makes soko2des.pl count how many of each character it processes
and output a list.  The main purpose of this is so that you can easily
get a count of how many "extra" boulders the level contains.  For
normal-sized Sokoban levels this is a very minor convenience.  For
large ones (like the Boulder Trophy Hall and the Metropolitan Museum
of Boulder Art) it is a quite major convenience.