Add two more variants of the first Sokoban level.
One of these is derived from a Slash'em Sokoban level, but I removed
the iron bars (on the grounds that the iron bars in Slash'em's Sokoban
are evil; invariably something with a ranged attack gets generated
behind them, and then you have to go past a bunch of times while
working on the puzzle, and you can't get in there to kill it until you
finish; on this particular level it's not nearly as bad as on some of
the others, but NetHack Fourk has a no iron bars in Sokoban policy on
general principle).  While I as editing, I also made some other
adjustments and gave the level a name ("Open the Door").

The other one ("Picking out the Seeds") is original.