Sokoban fixups:  fix Two-Phase so it can actually be completed; and augment the rings.
The previous changes to Two-Phase inadvertently disrupted the ability
to actually successefully complete it, so this fixes that.

Additionally, two of the new Sokoban prize levels are a little on the
chunky side, so add an extra random ring to each of them as a minor
compensation or for motivational purposes.