Correct the attribute-adding formula's handling of 18/** - 1.
Subtracting 1 from 18/** to get 18/99 for max strength, and adding 1
to 18 to get 19 max for some other attribute, isn't balanced, and it
isn't what I intended.  The problem was that there's overlap in the
way the game stores 18/** versus 19.  I'm not sure how the game keeps
them straight, but in any event, this formula handles the role-based
maximum adjustments in the intended fashion.

Whether the numbers for individual roles are as they should be remains
an open question.  But at least the formula is tested and working as
intended now.

This also makes the racial differences more significant for some
roles, as depending on race and role combo you can end up with a
maximum as low as 15 or as high as 21 for some attributes.