New playable race:  Sylph.
This is going to need some balance tweaks, but it's a starting point.

Sylphs are fairy creatures with some unusual abilities that come with
drawbacks.  The main one is that they regenerate health and energy by
drawing it from their environment, which they can do at a pretty good
rate, but this is blocked by wearing armor (other than a small
shield), by being in Gehennom, or by being on certain kinds of terrain
that are not amenable (lava, ice, air, ... basically anything that
prevents you from dipping your toes in; rock and water are both fine,
though) or levitating.

The tiles for the sylph monster (which _do_ get used for the player if
showrace is turned on) are not great, but until a real tiles artist
draws better ones, they'll have to do.