Remove redundant override, since get_gamedir does dumpdir handling now.
Originally, as per the comment in the removed section, get_gamedir did
handling for other cases but did not handle the dumps dir.  However,
such handling was added to get_gamedir as part of commit 94ff3f71ebd,
apparently in order to fix some Windows API-related issues; but at
that time the now-redundant handling was not removed from
init_game_paths in main.c.  This removes it.  Noticed by kerio when
doing the dgamelaunch install.