First set of demon lair changes: asbaal, yeedis
Instead of always getting the same version of Asmodeus' Lair and also
the same version of Baalzebub's lair, you now get, in any given game,
either Asmodeus' Lair *or* Baalzebub's, and also you get either
Yeenoghu's *or* Dispater's.  So you get the same total number of demon
lairs, but not always the same ones.

This implements just one version each of Yeenoghu's and Dispater's,
and it keeps the vanilla versions of Asmo's and Baalz's for now, even
though I hate both those lairs.  I did add one additional version of
Asmodeus' Lair, which is... mediocre, I guess.  I want to do better
ones, and more versions, but this is a start.