Shaped rooms, first draft.
This is my first attempt to allow shaped rooms.  Currently, it can do
L-shaped, T-shaped, S/Z-shaped ("Tetris"), and plus-shaped rooms.
Only rooms that are large enough have a chance to get shaped.

This isn't really ready for production yet.  For one thing, special
rooms, such as the Delphi room, can be shaped, and that shouldn't be.

Also, I haven't tested all the different orientations thoroughly, and
I haven't checked that the corners look right in utf8 mode.

At some point I want to add oval rooms and rooms with a pillar cut out
of the middle, and maybe also C shaped, maybe even I shaped, but those
would be future enhancements.

Also, I think I want the average cut size to be larger,
without changing the min or max.  (Maybe just use the max cut a
percentage of the time, not sure.)

But this is good enough to do a first commit.