Sylph playable race balance and flavor adjustments.
Instead of having intrinsic hunger, sylphs now burn nutrition
when they draw healing or power from their environment and
also when their healing ability fixes a status effect.  In
conjunction with this, they can't use the ability when weak
from hunger (or worse).

They can now wear any armor that is made of WOOD or CLOTH,
without interfering with their healing and energy drawing.
In practice this means small or elven shield, cloak, shirt,
and a fedora if they can find one.  (Cornuthaum or dunce cap
would also qualify, but those are probably more trouble than
they are worth.)

Sylphs now get an alignment hit for breaking vegetarian,
just like monks.  (It may occasionally be worth it, e.g.,
to eat a lizard corpse to stop stoning.)  Code has been
added to u_init.c to ensure that Sylphs (and while I was
at it, also Monks) will not be given non-vegetarian tins
in starting inventory.  (For Monks, this is currently not
relevant, as they aren't given random food anyway.)  The
reasoning is that tins, unlike most food, can't be used
to feed pets, and so if you can't eat it it's useless.
(Ideally, always-unsafe monsters like cockatrice shouldn't
be given in tins to any player at game start, but I have
not implemented that at this time, as it's not really a
Sylph-related issue.)

In addition to Priestesses, Healers, and (for now) Valks,
Sylphs can also be Rogues.