Shaped room bugfixes and enhancements, round 1.
I believe this fixes up the worst bugs in the shaped-room
implementation.  Foremost, it tries to avoid placing stuff
(monsters, objects, statues, traps, stairs, altars, whatever)
inside walls or solid rock.  Previously, the code just sort
of assumed that everything inside the room's rectangle was
room floor, more or less, and that's no longer the case.
As a side effect of how I implemented this, this, it also
probably means that objects and whatnot are less likely to
get placed on dungeon features (e.g., stairs), but that
seems like a largely positive side-effect on the whole.
However, there may be some cases that I missed, as there
are a lot of places in the code where things are placed
in rooms, and the x and y coordinates are almost always
picked separately, so I can't just change one function
and solve it.

I attempted to fix it so that special rooms (e.g., the
Delphi room, shops, etc.) don't get shaped.  This needs
further testing.

I also did some testing in UTF8 mode and fixed up the
appearance of the corner walls (TLCORNER vs BRCORNER vs
BLCORNER vs TRCORNER) for shaped rooms.  This is probably
also relevant for tiles players.

Also, when a shaped room is lit, the code now doesn't
light up blocks of stone within the room's rectangle that
are not part of the actual room due to shaping.  This is
minor, but it would be kinda weird to dig into solid rock
and have it be pre-lit.

I also added support for two new room shapes.  The one I am
calling "O" is a rectangle with a pillar cut out of the middle.
The other, which I am calling "X" for lack of a better name,
has all four corners cut out, like the plus-shaped room, but
also has the center cut out.  If this room shape gets the
maximum possible size of cut in both dimensions, there are
diagonal squeezes, so I will probably make this room shape
less common than the others when I get around to adjusting
the probabilities.

Oh, and I increased (slightly) the probability of a larger
cut, by introducing a chance for small cuts to get upgraded
to max size (decided independently in each dimension).

Oh, and I roughly halved the frequency of early eucalyptus,
because of the specially-placed early items, it's the least
important (except for healers, who get some anyway).

At this point there's some debugging code in there, although
it's #ifdeffed out.  Eventually I'll remove it entirely.