Add "variant" field to livelog and xlogfile, indicating "Fourk".
This is technically unnecessary, because people handling a logfile
*theoretically* should know where the file came from anyway.  But eh,
it's not like adding this field is hard to do or introduces any
significant maintenance burden.  Making it a separate field (rather
than changing the version field) avoids annoying people who expect
the logfile format to be stable, e.g., the cross-variant scoreboard
maintainer.