Mark shaped rooms as irregular and don't select them for shops.
Also, set the roomno field for their floor tiles so that somexy() can
cope with them.  This allows temple priests in shaped temples to know
which room they're supposed to be in, and I think it's also necessary
to allow branch portals/stairs to be placed in shaped rooms -- without
it, placebranch was giving me an impossible.

Not selecting shaped rooms for shops probably exerts a slight negative
pressure on average shop size (since small rooms cannot be shaped).
This, however, is almost certainly better than trying to make all the
shop code deal with shaped rooms.  In my testing, I still got a number
of large shops, and the shaped-room frequency is higher right now than
it ultimately will end up being (because the feature is new and I want
players to be able to find shaped rooms easily at first).  So on the
whole I think it's ok.  Especially given that I have already boosted
the odds of a general store in minetown.