Remove unused, partially-implemented rectangularize() function.
This function was created as one approach to solving the shaped shop
bug, but I ended up settling on a different solution to that, so this
function is not needed.  Since it was never completed and should not
be used, I'm removing it.  (I committed with it intact initially, to
get it into the git repo, so it's not being lost as such.  But it
isn't needed at this time.)