Call unseen monsters "something" rather than "it".
The misuse of the word "it" in this situation has bugged me for a long
time.  "It" tends to sound quite odd when it doesn't have an
antecedent, and when it _does_ seem to have one (in NetHack), it's
typically misleading, or occasionally just inadvertently silly.  The
thing is, English _has_ a perfectly cromulent pronoun that we can use
legitimately to refer to an unidentified entity.