Make nymphs blink away, rather than teleporting.
This adds support for monsters that blink away when they successfully
steal something (M3_BLINKAWAY), and nymphs are the first monsters to
get this new flag.  Nymphs no longer have M1_TELEPORT, so they don't
teleport under other circumstances, and their corpses don't grant
teleportitis.

Blinking away works on a straight-line clear-path basis, in increments
of 15 degrees, with a maximum radius dependent on monster level.  The
ability to actually /see/ the tile is not required, so nymphs can
blink into the dark or while blind.  All angles are tried (in a random
order) at the maximum distance first, and then if no valid
destinations are found the radius is backed off.  If it gets down to 2
with still no valid destinations (which could happen in tight spaces
or on a crowded or water-intensive level), the monster may end up
getting placed in a random location as a last resort.