Attempt to address the shaped temple bug.
When choosing a position for a shrine altar, still choose the fixed
center-of-the-room position if that's a normal room-floor tile; but
otherwise, attempt to find a better position.

What I've done is, for each tile in the room, count how many of the
tiles in the 3x3 area centered on it have the maximum available number
of floor tiles in the 3x3 area centered on them, and pick one of the
tiles that have the maximum value for this.  In most cases, this will
place the altar in an open area of the room, away from edges,
regardless of the room shape.

In a very large shaped room this will likely fail to get anywhere near
the middle; but most NetHack rooms are not so large that this is
likely to be a very big deal very often; and it's clearly better than
putting the altar (and priest if applicable) on solid rock tile in the
middle of a walled-off stone pillar that isn't considered part of the
room even (which made the priest abandon his post), as had been
observed happening without this fix.