Add UI option to accept implicit alternate menu accelerators (a, b, c, ...)
The option is off by default and is not recommended for new players.
It's only useful for people who have trained their muscle memory on
certain other variants.

If the option is set, you can press either the specified accelerator or
(if this does not conflict with a specified accelerator) you can press
the letter of the alphabet corresponding to how far down the list the
choice you want is -- a for the first one, b for the second, etc.

If there's a conflict, specified accelerators win.

In conjunction, I reordered the options on the call/name menu so that
(most of) their implicit alternate accelerators will match the other
variant in question.  There are probably other menus that need similar
treatment, but this is the menu for which the feature was requested.