Grimtooth buff:  poison attack, poison res, and anti-elf
Grimtooth is now permanently poisoned and provides poison resistance.

It also does extra damage to elves, even if they are poison resistant,
and provides warning of elves.

In order to implement all of this, some of the artifact special-attack
code has been adjusted so that an artifact can potentially have more
than one kind of attack bonus, e.g., it can have a bonus against a
certain kind of monster and also do elemental damage generally.

Also in order to make this work, I reintroduced the M2_ELF flag,
because that was easier than implementing warn of race.