Make backstab apply when attacking monsters with various status effects.
Rogues now get the backstab bonus when attacking monsters that are
fleeing, trapped, frozen, paralyzed or unable to move, sleeping,
blinded, confused, or stunned.

I attempted to make it also apply against peaceful monsters, but that
doesn't seem to work, probably because the mpeaceful flag is removed
when the monster is attacked, before the point in the code that
applies the backstab bonus is reached.  Perhaps that's just as well;
getting a bonus against peacefuls would've been rather a double-edged
sword (particularly if pets count as peaceful, which I think they do).