For the call/name menu only, allow a/b/c accelerators.
This works only for the first three items (a for C, b for y, c for n)
but works independently of the menu_use_abc setting (which will
consequently be deprecated and probably go away entirely soon).  It's
implemented using "group accelerators", which are intended for letting
you select multiple items with a category key, but whatever, it gets
the job done, and it's specific to the one menu.