Revert most of "Add UI option to accept implicit alternate menu accelerators (a, b, c, ...)"
This reverts the bulk of commit e574e0f335f94e2d935e078e343ed1583d1b46b6.

Not Reverted:
	libnethack/src/do_name.c

The reordering of the name/call menu is still wanted, because we
are still allowing a/b/c accelerators for that menu.  We're just
doing so via a different mechanism now (abusing group accelerators).