Oracle/Pacifist fix: Oracle always good for two experience levels.
Your first major and first minor consultation are now each always
good for one experience level gain, regardless of how much gold
you are charged, and also regardless of the order in which you do
them.  Pacifist is hard enough without my "make Oracles easier to
buy for newbies" fix inadvertently making it harder for Pacifists
to gain levels.  This fix technically makes it easier -- because
they spend less gold doing so -- but honestly, I doubt if gold
scarcity is likely to be a major factor in pacifist difficulty.