Don't attempt to add_damage on an undefined level, resulting in a segfault.
If add_damage() is called while level is null, the game segfaulted.
This commit preempts the segfault with an impossible() in that
instance, and it also stops the trap code (one level up the call
chain) from calling add_damage() in the first place if lev is not
the current level, since I believe all actual situations where the
player genuinely should be charged for shop damage can only occur
while lev and level are equal.

I am not yet sure whether this bug was exposed by changes in Fourk
(possibly one of my special levels somehow triggers it, although I do
not see how, since it fundamentally doesn't do anything really novel;
but since the specific level that was loading when the segfault
occurred is indeed unique to Fourk, this is a possibility; it is also
possible that some of my 4.3-maint work is responsible, by having made
some incorrect change to the level loading code), or whether it has
been dormant for some time (possibly from upstream, or conceivably
even from vanilla) and was merely exposed by the RNG's choice of trap
placement in Mandevil's game.  I may continue to analyze exactly where
the blame lies here.  But for now, this patch will allow the game in
question to be continued.

The player also claimed that prayer reliably segfaults in this game,
but I was unable to reproduce that.