Enhance a prayer-related impossible() message, in case someone gets it again.
A player filed an issue in the github web interface, which I later
discovered when another user filed another issue there and mentioned
it on IRC.  The first issue has to do with prayer, something other
players have also been complaining about ("prayer crashes"), but which
I have thus far not managed to reproduce, not for lack of trying; the
filed issue contained enough information to allow me to know which
impossible() was triggering, but not why.  This commit adds the
trouble number to the message, so that if it happens again I'll have
slightly more information to go on.