Intelligence-draining attacks no longer also cause partial amnesia.
Flayers are dangerous enough without the forget-items effect, and
it's annoying out of all proportion to how much it actually puts
the character at risk.  So it's gone.

I'm probably going to nuke amnesia altogether at some point, but the
scroll will need to be replaced, so that's on hold until I work out a
replacement for it.