Add an unaligned altar in Vlad's Tower.
This can be converted and used by the player.

The Minetown altar may have a cross-aligned priest, whom the
player may prefer to keep alive; and the Valley altar cannot
be converted because it is in Gehennom.  So here is a third
non-high altar that occurs in every game (barring bones).
This one is always available to be converted, once you can
reach it.