Fix the hilarious bones depth bug.
For a while, in Fourk, bones created on non-special levels in the main
dungeon at depth 20 and higher were saved as level-two bones and would
be loaded on DL2.  I'm so used to dealing with languages that have
string concatenation, I forgot to take into account that the code
for appending to a string actually needs to change if the string is
a different length.  This fixes the oversight.