Initial implementation of shield skill.
The details will probably need to be tweaked after testing.  Which
role gets which cap is open to reconsideration, and the numbers for
how the skill gets added into your AC formula are highly preliminary
at this point.  I'm also not sure the exercise criterion is right.

But I'm committing to get this out of my working directory and
into a code branch.  It can be revised later.