Fix monster level calculation on DL1 in challenge mode.
Someone had pointed out that the calcuations for minimum and maximum
monster level on DL1 were contradictory in challengemode.  This wasn't
crashing or anything (because of the way the minimum and maximum are
used individually rather than together by the code below, and because
the maximum is a soft check); but it lead to the actual range of
generated monsters being inconsistent with the min and max variables,
which at best is very unclear for code divers.  This commit explicitly
adjusts the maximum upward until it is strictly greater than the
minimum.