Allow wands of digging to dig multiple tiles in Gehennom.
The restriction to one tile for wands of digging in Gehennom was
probably intended for balance reasons (to prevent the mazes from being
subverted), but all it really does is make everyone carry a pick-axe
throughout Gehennom, stockpiling needlessly many wands for ascension
run, and limiting player carry capacity resulting in more stash trips,
which is boring but not actually challenging.

So I'm relaxing the restriction.  Wands of digging now work the
same way in Gehennom as in the main dungeon.