Try harder to avoid giving the player two of the same kind of book.
Thanks to Khor for reporting that his Sylph Priestess started with two
spellbooks of light (which also produced a weird error message about
already knowing the spell; the latter issue is not addressed yet in
this commit).  This was able to happen because the Priest role gets
two spellbooks in their base starting inventory list and then
sometimes rolls a third (which may be either light or protection); the
init code wasn't handling that case correctly.  In such cases, try a
bit harder to avoid duplication:  even if the type is specified, if
it's a spellbook or ring and the player is already carrying an
identical one, try to avoid duplicating it by picking a random object
of the same class instead.