Implement the msgnomerge option:  start each message on a new line.
This UI option was suggested by Khor.  With a sufficiently large value
for msgheight (which is easy to do on a modern system), it kind of
actually makes things more readable.  I may start using it myself.

However, it's turned off by default, because people who have a
small value for msgheight will presumably hate it.