Narrow the range for where Medusa's Island can appear.
Given that the location of Medusa's Island controls the size of the
"maze zone", the amount of variation in its depth was excessive.  I've
reduced it by two.  Normally I am a fan of variety, but in this case
effectively an entire section of the dungeon could be made almost
nonexistent, which does not seem ideal to me.

(People who hate mazes should bear in mind that the total number of
mazes in the dungeon will decrease very dramatically when the
aiscavern patch lands.  I do not plan to release a version with the
expanded post-Medusa maze zone unless it also has the ais caverns for
Gehennom.  So I'm not increasing the number of mazes.  I'm decreasing
the number of mazes and also adjusting their location somewhat.)