Use mkaiscav instead of makemaz for Gehennom filler levels.
The new algorithm, mkaiscav, is based on Perl code posted to rgrn by
ais523 a while back.  It generates a sort of cavern, but noticeably
different from Mines caverns, and with greater variety, taking the
dungeon depth into account in the generation algorithm so that the
early Gehennom levels are fairly open but later ones are somewhat more
broken up.

I've tweaked his algorithm somewhat and added liquids (rivers or pools
of water or lava) on some levels, as well as items and monsters.

Traps still need to be added, and further testing is needed.