Adjust the arrangement of Gehennom to clarify each special level's position.
On the whole, the order should be somewhat more defined now:  you get
Asmodeus or Baalzebub first, Juiblex after that, then Yeenoghu or
Dispater, which in turn is before the Wizard's Tower.

Orcus Town is an exception, at least for now:  it can still appear
either before the Yeenoghu or Dispater lair, or else below it.