Be more paranoid about feeding aisdepth to random-number functions.
The intention is for aiscavern levels to always have a positive
aisdepth, ranging from 1 to about 100 or so.  But the implementation
does to actually guarantee this, so when feeding it into functions
like rn2(), we have to armor it in a macro that checks for
non-positive values and replaces them with 1.  (In all cases where we
are doing something random based on depth below the valley, it is
theoretically correct to treat a depth of less than 1 as equal to 1.)