Choose location for placing a branch portal on aiscav levels.
Rather than relying on place_branch to choose a location (which
it is only programmed to do on maze levels and levels with rooms),
use aiscav-specific logic to choose a location pre-emptively.

This is an abstracted version of the same basic logic used to
place the up and down stairs -- and I intend to rework the stair
placement to use it in those cases as well, in a future commit,
to reduce code duplication.