Shaped room fixup:  don't cut both middle and corners if it will make a narrow squeeze.
If a room would have had both its middle and its corners cut out, but
this would have made a narrow squeeze, either the size of the cut is
reduced (if there's room), or else the corners are not cut after all,
leaving just a rectangular room with a pillar cut out.

Also, this commit cleans out some left-over debugging stuff that is
hopefully no longer needed.