Make Excalibur easier to get for knights, harder for others.
The way I did the formula, waiting to gain another level first will
alway improves your odds of getting Excalibur on the next dip, unless
you have already reached the point where it is guranteed.

Knights can now _reliably_ get Excalibur at XL5, without risk of
negative fountain effects, and can _potentially_ get it before XL5.

For most roles, the odds of getting Excalibur at XL5 are now half as
good as before, and it's not guaranteed until XL15 (by which time most
of the negative fountain effects are not so threatening anyway).

Lawful Barbarians, which are intended to be an easier role for new
players, can get it guaranteed at XL10, and their odds at XL5 are
similar to what they would be in vanilla.

As before, if you are not lawful, your longsword is cursed instead
of being converted to Excalibur, and made erodable if it wasn't.

This commit also adds the changelog entry for the recently merged
post-medusa and aiscavern changes.