Revise the Ranger quest goal level.
Jonadab's Comments:
The Ranger quest goal level in 3.4.3 was in need of adjustment.  In
the first place, it has nothing Ranger-themed about it, being a
holdover from the old Elf quest, wherein it was a goblin keep -- and
it only even _sort of_ made sense there.  Additionally, it was a
castle design, and the game has enough of those already.  Furthermore,
this particular castle design is more similar to some of the other
quest levels than is really comfortable.  So I pulled in this set of
changes from dnethack.  It's not perfect, but it's enough of an
improvement to stop the Ranger quest goal level from being a serious
contender for "worst quest level in the game", which it had otherwise
been in danger of since I merged Khor's revised ranger quest home.