Restore corpse and death-drop rates to vanilla (with 4.3-maint fix).
My changes (to make the probability of drops inversely proportional to
a logarithmic function of the number of monsters of the type that had
already been vanquished) were intended to improve game balance and
make farming (especially pudding farming) not seem compelling.  I am
not convinced that these changes worked as intended, so I am reverting
them for now.  I have a couple of ideas for how to approach the issue
from an entirely different angle.